flask-azuread-auth/requirements.txt
flask
python-dotenv
msal<2.0
